+++ src/classes/popup/edit/EditIconPopup.hpp
@@ -1,3 +1,4 @@
+#include "IconEditorState.hpp"
 #include <Geode/binding/SimplePlayer.hpp>
 #include <Geode/ui/Popup.hpp>
 #include <Geode/ui/TextInput.hpp>
@@ -13,21 +14,22 @@ protected:
     std::filesystem::path m_selectedPNG;
     std::filesystem::path m_selectedPlist;
     geode::Ref<cocos2d::CCArray> m_pages;
-    matjson::Value m_definitions;
+    IconEditorState m_state;
     geode::Ref<cocos2d::CCDictionary> m_pieces;
     geode::Ref<cocos2d::CCDictionary> m_frames;
     geode::Ref<SimplePlayer> m_player;
     geode::Ref<cocos2d::CCDictionary> m_sliders;
     geode::Ref<cocos2d::CCDictionary> m_inputs;
     std::string_view m_suffix;
-    cocos2d::CCArray* m_targets;
+    geode::Ref<cocos2d::CCDictionary> m_targets;
+    cocos2d::CCArray* m_targetsArray;
     cocos2d::CCSprite* m_selectSprite;
     cocos2d::CCMenu* m_pieceMenu;
+    cocos2d::CCSprite* m_mainColorSprite;
+    cocos2d::CCSprite* m_secondaryColorSprite;
+    cocos2d::CCSprite* m_glowColorSprite;
     IconType m_iconType;
     int m_page = 0;
-    int m_firstColor = 12;
-    int m_secondColor = 12;
-    int m_thirdColor = 12;
     bool m_hasChanged = false;
 
     bool setup(MoreIconsPopup* popup, IconType type) override;
@@ -35,8 +37,9 @@ protected:
     void updateControls(std::string_view id, float minimum, float maximum, float defaultValue, bool decimals);
     void transferPlayerToNode(cocos2d::CCNode* node, SimplePlayer* player);
     void addPieceButton(std::string_view suffix, int page, cocos2d::CCArray* targets);
-    void addColorButton(int& index, cocos2d::CCMenu* menu, const char* text, std::string id);
-    void updateWithSelectedFiles();
+    cocos2d::CCSprite* addColorButton(int& index, cocos2d::CCMenu* menu, const char* text, std::string id);
+    void updateColors();
+    bool updateWithSelectedFiles(std::string_view suffix = {});
     void updatePieces();
     void goToPage(int page);
     void updateTargets();