+++ src/classes/popup/edit/EditTrailPopup.hpp
@@ -14,6 +14,7 @@ protected:
 
     bool setup(MoreIconsPopup* popup) override;
     void addOrUpdateIcon(const std::string& name, const std::filesystem::path& path);
+    void updateWithPath(const std::filesystem::path& path);
     void saveTrail(const std::filesystem::path& path);
     void onClose(cocos2d::CCObject* sender) override;
 public: