+++ changelog.md
@@ -1,14 +1,16 @@
 # More Icons Changelog
 ## v1.16.0 (2025-12-06)
-- Added color previews and load/save state functionality to the icon editor
+- Added load/save state functionality to the icon editor
+- Added color previews to the icon editor
 - Added green flashing to the PNG/Plist buttons in the icon editor when the other file needs to be selected
 - Added pink selection to the PNG/Plist buttons in the icon editor when the file is selected
 - Changed the clear piece button to remove the extra piece instead of making it empty
 - Moved trail saving to the trail settings popup
 - Changed the icon viewer popup to show the icon name in the title
 - Fixed a bug where clicking a trail in the vanilla icon list popup would crash the game
-- Fixed a bug where saving an icon with a 0 size piece would crash the game
-- Fixed some pre-multiplied alpha issues when saving icons
+- Fixed a bug where saving an icon with a 0 size piece would crash the game (Reported by [Doritosxz](user:14012181))
+- Fixed a possible bug where the icon editor would not properly transform the extra piece preview
+- Fixed some pre-multiplied alpha issues when saving icons with presets
 - Removed some logs I left in by accident
 - Refactored the API in `MoreIconsV2.hpp`, utilizing Geode's event export system
   - This will replace `MoreIcons.hpp` in v2.0.0, which will be released after Geometry Dash v2.208